provide publish server method for zip-export
