+++ src/server.js
@@ -165,6 +165,7 @@ async function configure(options = { usePlugins: true, useOpenAPI: true }) {
     server.method('config', key => (key ? config[key] : config));
     server.method('generateToken', generateToken);
     server.method('logAction', require('@datawrapper/orm/utils/action').logAction);
+    server.method('createChartWebsite', require('./publish/create-chart-website.js'));
 
     const { validateThemeData } = schemas.initialize({
         getSchema: config.api.schemaBaseUrl
